// medico/Services/Medico.Api/ClientApp/src/app/patients/models/patient.ts
import { ZipCodeType } from './zipCodeType';

export class Patient {
    id: string;
    companyId: string;
    firstName: string;
    lastName: string;
    middleName: string;
    gender: number;
    dateOfBirth: any;
    maritalStatus: number;
    ssn: string;
    primaryAddress: string;
    secondaryAddress: string;
    city: string;
    primaryPhone: string;
    secondaryPhone: string;
    email: string;
    zip: string;
    state: number;
    patientInsuranceId: string;
    zipCodeType: ZipCodeType;
    notes: string;

    constructor() {
        this.zipCodeType = ZipCodeType.FiveDigit;
    }
}